Move picker into footer
